pyproject.toml
[project]
name = "reagt"
version = "0.1.0"
description = "ReAgt - Adaptive Training Agent for Endurance Athletes"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0",
    "fitdecode>=0.10",
    "rich>=14.0",
    "watchdog>=6.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that call external APIs (Gemini)",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
